feat - move Bank class
