+++ src/test/java/co/invest72/exchange_rate/domain/BankTest.java
@@ -1,4 +1,4 @@
-package co.invest72.money.domain;
+package co.invest72.exchange_rate.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import co.invest72.exchange_rate.infrastructure.api.FixedExchangeRateProvider;
+import co.invest72.money.domain.Currency;
+import co.invest72.money.domain.Money;
 
 class BankTest {
